Rename remote model.remote package to model.appcenter
