+++ src/main/java/io/jenkins/plugins/appcenter/api/AppCenterService.java
@@ -1,11 +1,11 @@
 package io.jenkins.plugins.appcenter.api;
 
-import io.jenkins.plugins.appcenter.model.remote.ReleaseDetailsUpdateRequest;
-import io.jenkins.plugins.appcenter.model.remote.ReleaseDetailsUpdateResponse;
-import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadBeginRequest;
-import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadBeginResponse;
-import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadEndRequest;
-import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadEndResponse;
+import io.jenkins.plugins.appcenter.model.appcenter.ReleaseDetailsUpdateRequest;
+import io.jenkins.plugins.appcenter.model.appcenter.ReleaseDetailsUpdateResponse;
+import io.jenkins.plugins.appcenter.model.appcenter.ReleaseUploadBeginRequest;
+import io.jenkins.plugins.appcenter.model.appcenter.ReleaseUploadBeginResponse;
+import io.jenkins.plugins.appcenter.model.appcenter.ReleaseUploadEndRequest;
+import io.jenkins.plugins.appcenter.model.appcenter.ReleaseUploadEndResponse;
 import retrofit2.http.Body;
 import retrofit2.http.PATCH;
 import retrofit2.http.POST;
