+++ src/main/java/io/jenkins/plugins/appcenter/model/appcenter/DestinationId.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model.remote;
+package io.jenkins.plugins.appcenter.model.appcenter;
 
 public final class DestinationId {
     public final String name;
