+++ src/main/java/io/jenkins/plugins/appcenter/model/appcenter/ReleaseDetailsUpdateResponse.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model.remote;
+package io.jenkins.plugins.appcenter.model.appcenter;
 
 public final class ReleaseDetailsUpdateResponse {
     public final String release_notes;
