+++ src/main/java/io/jenkins/plugins/appcenter/model/appcenter/Status.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model.remote;
+package io.jenkins.plugins.appcenter.model.appcenter;
 
 public enum Status {
     committed,
